Small refactor - moving general purpose methods to models.fhir from intervention_strategies.
